tests/wpt/web-platform-tests/animation-worklet/worklet-animation-pause.https.html

<!DOCTYPE html>
<title>Verify that currentTime and playState are correct when animation is paused</title>
<link rel="help" href="https://drafts.css-houdini.org/css-animationworklet/">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/web-animations/testcommon.js"></script>
<script src="common.js"></script>

<div id="box"></div>

<script>

setup(setupAndRegisterTests, {explicit_done: true});

function createAnimation() {
  const box = document.getElementById('box');
  const effect = new KeyframeEffect(box,
    { transform: ['translateY(100px)', 'translateY(200px)'] },
    { duration: 100, iterations: 1 }
  );

  return new WorkletAnimation('passthrough', effect);
}

async function setupAndRegisterTests() {
  await registerPassthroughAnimator();

  promise_test(async t => {
    const animation = createAnimation();
    animation.play();
    // Immediately pausing animation should freeze the current time at 0.
    animation.pause();
    assert_equals(animation.currentTime, 0);
    assert_equals(animation.playState, "paused");
    // Wait some time to ensure a paused animation actually freezes.
    await waitForNextFrame();
    assert_equals(animation.currentTime, 0);
    assert_equals(animation.playState, "paused");
  }, 'pausing an animation freezes its current time');

  promise_test(async t => {
    const animation = createAnimation();
    const startTime = document.timeline.currentTime;
    animation.pause();
    animation.play();
    await waitForNextFrame();
    const timelineTime = document.timeline.currentTime;
    assert_equals(animation.playState, "running");
    assert_greater_than(animation.currentTime, 0);
    assert_times_equal(animation.currentTime, (timelineTime - startTime));
  }, 'playing a paused animation should resume it');

  done();
}

</script>
